feat!: migrate to frontend-utils, drop transformer related code
# Conflicts:
#	CHANGELOG.md
#	package-lock.json
#	package.json
#	src/index.ts
