+++ src/constants/single-indentation.ts
@@ -1,0 +1,1 @@
+export const SINGLE_INDENTATION = "  ";
