--- src/formatter/index.ts
@@ -1,7 +1,0 @@
-import { indentDSL } from "./indent-dsl";
-
-const formatter = {
-  indentDSL,
-};
-
-export default formatter;
