--- src/parser/index.ts
@@ -1,10 +1,0 @@
-export { parseDSL } from "./parse-dsl";
-
-export {
-  ParserResult,
-  RelationDefOperator,
-  RelationDefParserResult,
-  RelationTargetParserResult,
-  RewriteType,
-  TransformedType,
-} from "./parser.typings";
