+++ src/theme/index.ts
@@ -2,7 +2,7 @@ import { openfgaDark } from "./supported-themes/openfga-dark";
 import { OpenFgaThemeConfiguration, SupportedTheme } from "./theme.typings";
 
 export * from "./theme.typings";
-export { languageDefinition } from "../syntax-highlighters/prism/language-definition";
+export { languageDefinition } from "../tools/prism/language-definition";
 
 export const supportedThemes: Record<SupportedTheme, OpenFgaThemeConfiguration> = {
   [SupportedTheme.OpenFgaDark]: openfgaDark,
