+++ src/validator/index.ts
@@ -1,9 +1,6 @@
-import { checkDSL } from "./check-dsl";
-export { ValidationOptions, ValidationRegex } from "./check-dsl";
 import { ulidValidate } from "./ulid-regex";
 
 const validator = {
-  checkDSL,
   ulidValidate,
 };
